trv_local_automation/requirements.txt
paho-mqtt==1.6.1
vboxapi==1.0
